# src/api/docs/swagger-old2.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "G-Taxi API",
    "description": "G-Taxi apis"
  },
  "basePath": "/",
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/users/createProfile": {
      "post": {
        "tags": [
          "Athletes"
        ],
        "summary": "Create user profile",
        "description": "Create user profile",
        "operationId": "createProfile",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "$ref": "#/parameters/HeaderAuth"
          },
          {
            "in": "formData",
            "name": "name",
            "value": "John Dae",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "school",
            "value": "London Public School",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "sport",
            "required": true,
            "value": "62a8157ccdb5e22870f60b9a,62a816d2cdb5e22870f60c42"
          },
          {
            "in": "formData",
            "name": "description",
            "value": "Some profile description",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "gender",
            "value": "Male/Female",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "position",
            "value": "position",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "city",
            "value": "New York",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "zip",
            "value": "AL3 8QE",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "state",
            "value": "Texas",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "rostarPicture",
            "value": "rostarPicture.xyz/img.jpg",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "avatar",
            "type": "file",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Create user profile",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
    "/users/createProfileTrainer": {
      "post": {
        "tags": [
          "Trainers"
        ],
        "summary": "Create user profile",
        "description": "Create user profile",
        "operationId": "createProfile",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "$ref": "#/parameters/HeaderAuth"
          },
          {
            "in": "formData",
            "name": "name",
            "value": "John Dae",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "school",
            "value": "London Public School",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "sport",
            "required": true,
            "value": "62a8157ccdb5e22870f60b9a"
          },
          {
            "in": "formData",
            "name": "division",
            "required": true,
            "value": "62a85512444e692570b7da4b"
          },
          {
            "in": "formData",
            "name": "description",
            "value": "Some profile description",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "gender",
            "value": "Male/Female",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "city",
            "value": "New York",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "zip",
            "value": "AL3 8QE",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "state",
            "value": "Texas",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "rostarPicture",
            "value": "rostarPicture.xyz/img.jpg",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "avatar",
            "type": "file",
            "required": true
          },
          {
            "in": "formData",
            "name": "hourly_price",
            "value": "100",
            "type": "number",
            "required": true
          },
          {
            "in": "formData",
            "name": "age",
            "value": "25",
            "type": "number",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Create Traner's profile",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
...
    "/users/homescreen": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "User's Homescreen",
        "description": "User's homescreen contains posts ",
        "operationId": "homescreen",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "$ref": "#/parameters/HeaderAuth"
          },
          {
            "in": "query",
            "name": "search",
            "type":"string"
          }
        ],
        "responses": {
          "200": {
            "description": "Posts on homescreen",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
...
    "/practice/password": {
      "put": {
        "tags": [
          "Practice"
        ],
        "summary": "Change password",
        "description": "Change passord",
        "operationId": "password",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "$ref": "#/parameters/HeaderAuth"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/change_password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Change password in system",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
...
    "/concierges/profile": {
      "get": {
        "tags": [
          "Concierges"
        ],
        "summary": "Get Concierge profile",
        "description": "Get Concierge profile",
        "operationId": "profile",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "$ref": "#/parameters/HeaderAuth"
          }
        ],
        "responses": {
          "200": {
            "description": "Get concierge profile",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "change_password": {
      "type": "object",
      "required": [
        "currentPassword",
        "newPassword",
        "cnfNewPassword"
      ],
      "properties": {
        "currentPassword": {
          "type": "string",
          "description": "Current Password",
          "format": "password",
          "example": "admin@1234"
        },
        "newPassword": {
          "type": "string",
          "description": "New Password",
          "format": "password",
          "example": "admin@12345"
        },
        "cnfNewPassword": {
          "type": "string",
          "description": "Confirm new Password",
          "format": "password",
          "example": "admin@12345"
        }
      }
    },
...
    "responsesOk": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "msg": {
          "type": "string"
        },
        "data": {
          "type": "object"
        }
      }
    },
    "responsesError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "msg": {
          "type": "string"
        },
        "error": {
          "type": "object"
        }
      }
    },
...
  },
  "parameters": {
    "HeaderVersion": {
      "name": "X-G-Taxi-Version",
      "in": "header",
      "description": "X-G-Taxi-Version",
      "required": true,
      "type": "string",
      "default": "1.0.0"
    },
    "HeaderLanguage": {
      "name": "Accept-Language",
      "in": "header",
      "description": "Accept-Language",
      "required": true,
      "type": "string",
      "default": "en"
    },
    "HeaderPlatform": {
      "name": "X-G-Taxi-Platform",
      "in": "header",
      "description": "android or ios or web",
      "required": true,
      "type": "string",
      "default": "android"
    },
    "HeaderAuth": {
      "name": "Authorization",
      "in": "header",
      "description": "Authorization",
      "required": true,
      "type": "string"
    },
...
  }
}